# Conflicts:
#	src/main/resources/templates/작업.txt
공지사항 첨부파일 기능 완료

+++ src/main/java/kwang/ho/mapper/board/AttachMapper.java
@@ -10,7 +10,7 @@ public interface AttachMapper {
 
     int modifyAttach(List<AttachDTO> attachList);
 
-    AttachDTO selectAttachDetail(Long idx);
+    AttachDTO selectAttachDetail(int idx);
 
     int deleteAttach(int boardIdx);
